fix(rust): resolve clippy findings, incl. two correctness bugs
Clean up every clippy lint across the client logic crates ahead of wiring
a CI clippy gate. Two of these were genuine correctness bugs, not style:

- login_probe.rs: `find(|m| m.connection == peer || true)` — the `|| true`
  made the connection filter dead (always matched the first message).
  Single-connection probe so behaviour was unchanged in practice, but the
  predicate was a bug. (clippy::overly_complex_bool_expr, deny-by-default)
- weather.rs: particle phase used the literal `6.283` for a full turn;
  replaced with `std::f32::consts::TAU` (clippy::approx_constant,
  deny-by-default) — slightly more accurate, functionally identical.

The rest are clarity/idiom, all semantics-preserving:
- actors.rs: drop redundant `as u8` on `read_byte()` (already u8).
- texture.rs: `rsplit(['\','/'])` array pattern; remove no-op `* 1`.
- bloom.rs: `var_os(..)?` guard; drop `.clone()` on Copy BindGroupLayoutEntry.
- assets.rs / client_window.rs: `(a..b).contains(&x)` for manual range checks.
- client_window.rs: trim float literals to f32 precision (identical bits;
  HUD button x-fractions unchanged) and iterate `bar_ids.iter().enumerate()`
  instead of indexing a `0..12` range.
- trade_test.rs: break a doc list so the trailing line isn't a lazy
  continuation.

+++ client-rs/crates/rcce-data/src/actors.rs
@@ -151,7 +151,7 @@ fn parse_record(r: &mut BlitzReader) -> Result<ActorTemplate, ReadError> {
     for _ in 0..20 {
         r.read_short()?;
     }
-    let genders = r.read_byte()? as u8;
+    let genders = r.read_byte()?;
     let playable = r.read_byte()? != 0;
     let _rideable = r.read_byte()?;
     let _aggressiveness = r.read_byte()?;
